DROID-3347 refactoring
